+++ shared/src/main/java/net/blay09/mods/inventoryessentials/mixin/CreativeModeInventoryScreenAccessor.java
@@ -1,0 +1,12 @@
+package net.blay09.mods.inventoryessentials.mixin;
+
+import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
+import net.minecraft.world.SimpleContainer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(CreativeModeInventoryScreen.class)
+public interface CreativeModeInventoryScreenAccessor {
+    @Accessor
+    SimpleContainer getCONTAINER();
+}