+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-### Version 5.0.0 for Minecraft 1.19.x
-
-- Updated to Minecraft 1.19
+- Fixed weird behaviour when shift-clicking in the creative mode screen while dragging the mouse